doc-convert/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# 核心依赖：先安装 numpy
numpy==1.24.3

# PDF 转换
PyMuPDF==1.24.3
pdf2docx==0.5.8
img2pdf==0.5.1

# Office 文档处理
python-docx==1.1.2
python-pptx==0.6.23

# 图像处理
Pillow==10.3.0
pillow-heif==0.16.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10

# 工具库
aiofiles==23.2.1
